fix: trying to implement a neovim test fixture
Not easy. I was planning to have neovim communicate
with the test fixture via a FIFO. But Bazel does not
like FIFOs, because they are not hermetic.

So I now have to go back to the drawing board,
and figure out how to have a remote controlled
neovim that actually works under bazel.