fix: allows removing the singleton annotation
Fixes a bug in which diagnostics are not updated when
a single annotation is deleted in a file.

The old code skipped sending diagnostics update when there
are no annotations. However, in the case when a singleton
annotation is removed, we still need to update diagnostics
so that it can be removed in the editor.

Fixes: #5